Moved interface under namespace Contract
